+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Query History fix appending after overflow
 - New entity editor 🎉
 - Add from URL autofocus on input on dialog open
+- Allow display of empty query history
 
 ## [2021.05.13]
 - Double-click on graph view to fit content